- added early stopping - summary metrics - updated loss function to use pdf and sigma loss per end-state - bugfix for cases where float64 was being forced - bugfix for plotting vertical lines when density plots fails - plot 5th and 95th lines/area along with mean - added example notebook demonstrating API use - bumped to v0.0.4

+++ inverse_canopy/params.py
@@ -15,8 +15,8 @@ class ModelOutputs(dict):
         self['means'] = means
         self['stds'] = stds
 
-    def to_musigma(self) -> 'ModelParams':
-        mus, sigmas = compute_mu_sigma(self['means'], self['stds'])
+    def to_musigma(self, dtype=tf.float64) -> 'ModelParams':
+        mus, sigmas = compute_mu_sigma(self['means'], self['stds'], dtype=dtype)
         return ModelParams(mus=mus, sigmas=sigmas)
 
     def __repr__(self):
@@ -60,8 +60,8 @@ class ModelParams(dict):
     def spread(self):
         return self['mus'], self['sigmas']
 
-    def to_meanstd(self) -> ModelOutputs:
-        means, stds = compute_mean_std(self['mus'], self['sigmas'])
+    def to_meanstd(self, dtype=tf.float64) -> ModelOutputs:
+        means, stds = compute_mean_std(self['mus'], self['sigmas'], dtype=dtype)
         return ModelOutputs(means=means, stds=stds)
 
     def __repr__(self):
